Move pages to components
