+++ containers/vuejs/src/components/HomePage.vue
