--- containers/vuejs/src/router/pages/ProfilePage.vue
@@ -1,13 +1,0 @@
-<template>
-	<main id="Profile">
-		<h1>Profile</h1>
-		<p>This is the profile page This is the profile page This is the profile page This is the profile page This is the
-			profile page This is the profile page</p>
-	</main>
-</template>
-
-<style>
-h1 {text-align: center;}
-p {text-align: center;}
-div {text-align: center;}
-</style>
